feat: 支持local_config启动 add: 增加对crt path的说明，为传入config.py的相对路径 move: web component
