+++ unilabos/app/web/utils/device_utils.py
