+++ unilabos/app/web/utils/ros_utils.py
@@ -7,7 +7,7 @@ ROS 工具函数模块
 import traceback
 from typing import Dict, Any
 
-from unilabos.web.utils.action_utils import get_action_info
+from unilabos.app.web.utils.action_utils import get_action_info
 
 # 存储 ROS 节点信息的全局变量
 ros_node_info = {"online_devices": {}, "device_topics": {}, "device_actions": {}}
